kafka-producer/requirements.txt
kafka-python
psutil